remove conversion of loss to float before scaling (tests failing)

+++ deepspeed/pt/deepspeed_light.py
@@ -703,7 +703,7 @@ class DeepSpeedLight(Module):
 
         # scale loss w.r.t. gradient accumulation if needed
         if self.gradient_accumulation_steps() > 1:
-            loss = self._scale_loss(loss.float())
+            loss = self._scale_loss(loss)
 
         # Log training Loss
         if self.tensorboard_enabled():
